requirements.txt
TwitterAPI==2.2.3
coverage==3.7.1
mock==1.0.1
nose==1.3.4
oauthlib==0.6.3
python-dateutil==2.2
pytz==2014.7
requests==2.4.1
requests-oauthlib==0.4.1
simplejson==3.6.3
six==1.8.0
wsgiref==0.1.2
